Create basic API infrastructure and authentication/order endpoints

+++ config/notifications.php
@@ -3,5 +3,5 @@
 return [
     'default_channel' => env('NOTIFY_DEFAULT_CHANNEL', 'sms'),
     'fallback_channel' => 'email',
-    'dispatch_batch' => env('NOTIFY_DISPATCH_BATCH', 50),
+    'dispatch_batch' => (int) env('NOTIFY_DISPATCH_BATCH', 50),
 ];
